+++ homeassistant/components/tahoma_refactor/switch.py
@@ -4,8 +4,8 @@ import logging
 from homeassistant.components.switch import SwitchEntity
 from homeassistant.const import STATE_OFF, STATE_ON
 
-from . import TahomaDevice
 from .const import DOMAIN, TAHOMA_TYPES
+from .tahoma_device import TahomaDevice
 
 _LOGGER = logging.getLogger(__name__)
